feat(META): migrate to a real Bundler
