+++ jsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "target": "es2018",
+    "jsx": "react",
+    "checkJs": true
+  }
+}
