+++ now.json
@@ -6,7 +6,7 @@
       "src": "package.json",
       "use": "@now/static-build",
       "config": {
-        "distDir": "dist"
+        "distDir": "public"
       }
     }
   ]
