+++ src/index.html
@@ -4,11 +4,10 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Davecarter.dev => Personal site</title>
-  <link rel="manifest" href="./manifest.webmanifest">
+  <link rel="manifest" href="/static/manifest.webmanifest">
   <link href="https://fonts.googleapis.com/css?family=Fira+Sans" rel="stylesheet">
 </head>
 <body>
   <div id="app"></div>
-  <script src="./src/index.js"></script>
 </body>
 </html>
