+++ src/index.scss
@@ -17,7 +17,7 @@ h1, h2, h3, h4, h5, h6 {
 }
 
 main {
-  background: url('./images/bg_heading.jpg') center center no-repeat;
+  background: url('/static/images/bg_heading.jpg') center center no-repeat;
   background-size: cover;
   height: 100vh;
 }
